category's system

+++ database/migrations/2019_11_29_181419_create_sellers_table.php
@@ -35,7 +35,7 @@ class CreateSellersTable extends Migration
         $sellers = Seller::all();
         foreach($sellers as $seller)
         {
-            unlink(storage_path("app\public/{$seller->image}"));
+            if(!empty($seller->image)) unlink(storage_path("app\public/{$seller->image}"));
         }
         Schema::dropIfExists('sellers');
     }
